training/requirements.txt
PyYaml
protobuf
python-chess
pytorch-lightning
torch
torchsummary